BackEnd/backend/requirements.txt
django>=4.0.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7
pymongo>=4.6.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
rsa>=4.9
requests>=2.31.0